# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autonomous Recruiting Agent Swarm",
    "description": "AI-powered multi-agent system for recruiting",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List Jobs",
        "description": "List all jobs",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create Job",
        "description": "Create a new recruiting job.\n\nUses content-based duplicate detection: if a job with identical content already exists,\nreturns the existing job instead of creating a duplicate. This prevents multiple job_ids\nfor the same logical job regardless of when it was created.",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Job",
        "description": "Get job details by ID",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/start": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Start Job",
        "description": "Start the agent pipeline for a job",
        "operationId": "start_job_api_jobs__job_id__start_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/find-more": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Find More Candidates",
        "description": "Find more candidates for an existing job.\n\nThis endpoint re-runs the agent pipeline but excludes already-found candidates,\nuses pagination to get different results, and randomizes search strategies.",
        "operationId": "find_more_candidates_api_jobs__job_id__find_more_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "List Candidates",
        "description": "List candidates, optionally filtered by job_id",
        "operationId": "list_candidates_api_candidates_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candidate"
                  },
                  "title": "Response List Candidates Api Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "Get Candidate",
        "description": "Get candidate details",
        "operationId": "get_candidate_api_candidates__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/message": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "Get Candidate Message",
        "description": "Get outreach message for a candidate",
        "operationId": "get_candidate_message_api_candidates__candidate_id__message_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/candidates": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search Candidates By Strengths",
        "description": "Search candidates using semantic similarity across ALL jobs.\n\nThis endpoint searches the entire candidate pool globally, not filtered by job.\nA candidate who applied for one position might be perfect for another based on\ntheir strengths and skills.\n\nArgs:\n    query: Natural language query describing desired candidate strengths (min 3 chars)\n    limit: Maximum number of results (default 10, max 100)\n\nReturns:\n    List of candidates ranked by similarity across all jobs",
        "operationId": "search_candidates_by_strengths_api_search_candidates_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vector/candidates": {
      "get": {
        "tags": [
          "Vector"
        ],
        "summary": "Get Vector Candidates",
        "description": "Get all candidates for a job from Weaviate vector database.\n\nArgs:\n    job_id: Job ID to filter by (required)\n    min_fit_score: Minimum fit score threshold (optional)\n\nReturns:\n    List of candidates from vector database",
        "operationId": "get_vector_candidates_api_vector_candidates_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "min_fit_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Min Fit Score"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Job": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "title": "Description"
          },
          "requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirements"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "company_name": {
            "type": "string",
            "minLength": 1,
            "title": "Company Name"
          },
          "company_highlights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Company Highlights"
          },
          "model_provider": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(claude|gemini)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Provider"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "company_name",
          "id",
          "created_at",
          "status"
        ],
        "title": "Job",
        "description": "Schema for job response"
      },
      "JobCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "title": "Description"
          },
          "requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirements"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "company_name": {
            "type": "string",
            "minLength": 1,
            "title": "Company Name"
          },
          "company_highlights": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Company Highlights"
          },
          "model_provider": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(claude|gemini)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Provider"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "company_name"
        ],
        "title": "JobCreate",
        "description": "Schema for creating a new job"
      },
      "JobStartResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "message",
          "job_id",
          "status"
        ],
        "title": "JobStartResponse",
        "description": "Response for starting a job"
      },
...
    }
  }
}